Override default thunderbird desktop file
The override makes sure the birdtray application is started together with
thunderbird. This is the only way I found that allows starting birdtray
semi-automatically. When tried with XDG autostart or exec-once in hyprland
config, it always was giving:
```
Sorry, the system tray cannot be controlled by this add-on on your operating
system.
```